FindClippingBase in lib-builtin-effects
